Installed react icons, moved component folder from assets to src, worked on the header page and the styling
