+++ src/App.jsx
@@ -1,10 +1,12 @@
 import "./App.css";
+import Header from "./Pages/Header";
 import Router from "./Routes/Router";
 
 const App = () => {
   return (
     <div>
       <Router />
+      <Header/>
     </div>
   );
 };
