+++ src/Layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import Header from "../components/Header"; 
+import Header from "../Pages/Header.jsx"; 
 
 const Layout = () => {
   return (
