+++ src/Pages/HeaderStyle.css
@@ -1,0 +1,95 @@
+.emeka_header {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
+  background-color: #ffffff;
+}
+.emeka_header_left {
+  /* background-color: red; */
+  display: flex;
+  align-items: center;
+  gap: 30px;
+}
+.emeka_header_logo {
+  width: 80px;
+  height: auto;
+}
+.location {
+  display: flex;
+  align-items: center;
+  font-size: 16px;
+  gap: 10px;
+}
+.emeka_location {
+  display: flex;
+  align-items: center;
+  font-size: 22px;
+  color: orange;
+}
+.emeka_dropdown{
+  font-size: 25px;
+  cursor: pointer;
+  transition: transform 0.2s ease;
+  color: orange;
+  display: flex;
+  align-items: center;
+}
+.emeka_header_right {
+  display: flex;
+  align-items: center;
+  gap: 30px;
+  /* background-color: blue; */
+}
+.search_container {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+.search_icon {
+  position: absolute;
+  left: 10px; 
+  color: gray;
+  font-size: 16px;
+}
+.emeka_search_bar {
+  width: 250px; 
+  padding: 8px 10px 8px 35px;
+  background-color: lightgrey;
+  border: none;
+  outline: none;
+  border-radius: 6px;
+  /* outline: 1px solid orange;
+  border: 1px solid orange; */
+  /* border-radius: 8px; */
+}
+.emeka_search_bar:hover{
+    border: 1px solid orange;
+    outline: 1px solid orange;
+}
+.emeka_cart_container {
+  position: relative;
+  display: inline-flex;
+  align-items: center;
+  cursor: pointer;
+}
+.cart_count{
+  position: absolute;
+  top: -6px;
+  right: -10px;
+  background-color: orange;
+  color: white;
+  font-size: 11px;
+  padding: 2px 5px;
+  border-radius: 50%;
+  min-width: 18px;
+  text-align: center;
+}
+.emeka_icon {
+  font-size: 20px;
+  cursor: pointer;
+  color: orange; 
+}
+
+
